Add timestamp to filenames of APKs built on CI
With the index-v2 format of F-Droid fdroidserver assumes that APKs of
different versions of an app have different filenames. If the APKs have
the same name then F-Droid will keep showing the date when an APK filename
was seen for the first time as "Added at" date in the version list.
Using different filenames ensures that fdroidserver shows the correct
"Added at" dates.